third_party/blink/web_tests/external/wpt/payment-method-basic-card/payment-request-canmakepayment-method.https-expected.txt

This is a testharness.js-based test.
FAIL Must return false when the PMI is not supported at by the user agent. promise_test: Unhandled rejection with value: object "UnknownError: Renderer process could not establish or lost IPC connection to the PaymentRequest service in the browser process."
FAIL Must return true when basic-card is amongst unsupported PMIs. promise_test: Unhandled rejection with value: object "UnknownError: Renderer process could not establish or lost IPC connection to the PaymentRequest service in the browser process."
FAIL If basic-card is supported, then return a promise that resolves to true. assert_equals: if it throws, then it must be a NotAllowedError. expected "NotAllowedError" but got "UnknownError"
Harness: the test ran to completion.
